+++ src/note/metadata.rs
@@ -6,7 +6,7 @@ use std::collections::HashMap;
 
 #[serde_as]
 #[skip_serializing_none]
-#[derive(Deserialize, Serialize, Debug)]
+#[derive(Default, PartialEq, Serialize, Deserialize, Debug)]
 pub struct Metadata {
     pub title: Option<String>,
     pub description: Option<String>,
@@ -18,12 +18,12 @@ pub struct Metadata {
 
     #[serde(flatten)]
     #[serde_as(as = "HashMap<DisplayFromStr, _>")]
-    others: HashMap<String, serde_yaml::Value>,
+    pub others: HashMap<String, serde_yaml::Value>,
 }
 
 #[serde_as]
 #[skip_serializing_none]
-#[derive(Serialize, Deserialize, Debug)]
+#[derive(Default, PartialEq, Serialize, Deserialize, Debug)]
 #[serde(rename_all = "camelCase")]
 pub struct Bookmark {
     pub id: Option<BookmarkId>,
@@ -35,7 +35,7 @@ pub struct Bookmark {
     others: HashMap<String, serde_yaml::Value>,
 }
 
-#[derive(Serialize, Deserialize, Debug)]
+#[derive(Default, PartialEq, Serialize, Deserialize, Debug)]
 pub struct BookmarkId(String);
 
 impl Metadata {
