+++ src/note/model.rs
@@ -2,14 +2,14 @@ use markdown::mdast::Node;
 
 use crate::Metadata;
 
-#[derive(Debug)]
+#[derive(PartialEq, Debug)]
 pub struct Note {
     pub metadata: Option<Metadata>,
     pub summary: Option<Block>,
     pub blocks: Vec<Block>,
 }
 
-#[derive(Debug)]
+#[derive(PartialEq, Debug)]
 pub enum Block {
     Empty,
     Container {
@@ -22,7 +22,7 @@ pub enum Block {
     Node(Node),
 }
 
-#[derive(Debug)]
+#[derive(PartialEq, Debug)]
 pub enum CardKind {
     Note,
     Summary,
